pygl/requirements.txt
PyOpenGL>=3.1.7
numpy>=2.1.3
glfw>=2.6.5
PyGLM>=2.7.1
pillow>=11.0.0
